refactoring to get better perspective
