+++ lib/easy_serializer/base.rb
@@ -47,10 +47,10 @@ module EasySerializer
     alias_method :to_hash, :serialize
     alias_method :to_s, :to_json
 
-    def send_to_serializer(serializer, value)
-      return unless value
-      option_to_value(serializer, value).call(value)
-    end
+    # def send_to_serializer(serializer, value)
+    #   return unless value
+    #   option_to_value(serializer, value).call(value)
+    # end
 
     private
 
@@ -81,45 +81,49 @@ module EasySerializer
     end
 
     def value_or_default(setup)
-      value = attr_serializer(setup)
-      if value.nil? && setup[:default]
-        return option_to_value(setup[:default], object)
-      end
-      value
+      Attribute.call(self, setup)
     end
 
-    def attr_serializer(setup)
-      value = cache_or_attribute(setup)
-      return value if value.respond_to?(:fetch) || value.respond_to?(:each)
-      return value unless serializer = setup[:serializer]
-      if setup[:collection]
-        Array.wrap(value).map { |o| cache_or_serialize(serializer, o, setup) }
-      else
-        cache_or_serialize(serializer, value, setup)
-      end
-    end
-
-    def cache_or_attribute(setup)
-      execute = setup[:block] || proc { |o| o.send(setup[:name]) }
-      if EasySerializer.perform_caching && setup[:cache]
-        Cacher.call(self, setup, nil, &execute)
-      else
-        if object.respond_to?(:each)
-          object.map { |o| instance_exec o, &execute }
-        else
-          instance_exec object, &execute
-        end
-      end
-    end
-
-    def cache_or_serialize(serializer, value, opts)
-      return unless value
-      if EasySerializer.perform_caching && opts[:cache]
-        Cacher.call(self, opts, value)
-      else
-        send_to_serializer(serializer, value)
-      end
-    end
+    # def value_or_default(setup)
+    #   value = attr_serializer(setup)
+    #   if value.nil? && setup[:default]
+    #     return option_to_value(setup[:default], object)
+    #   end
+    #   value
+    # end
+    #
+    # def attr_serializer(setup)
+    #   value = cache_or_attribute(setup)
+    #   return value if value.respond_to?(:fetch) || value.respond_to?(:each)
+    #   return value unless serializer = setup[:serializer]
+    #   if setup[:collection]
+    #     Array.wrap(value).map { |o| cache_or_serialize(serializer, o, setup) }
+    #   else
+    #     cache_or_serialize(serializer, value, setup)
+    #   end
+    # end
+    #
+    # def cache_or_attribute(setup)
+    #   execute = setup[:block] || proc { |o| o.send(setup[:name]) }
+    #   if EasySerializer.perform_caching && setup[:cache]
+    #     Cacher.call(self, setup, nil, &execute)
+    #   else
+    #     if object.respond_to?(:each)
+    #       object.map { |o| instance_exec o, &execute }
+    #     else
+    #       instance_exec object, &execute
+    #     end
+    #   end
+    # end
+    #
+    # def cache_or_serialize(serializer, value, opts)
+    #   return unless value
+    #   if EasySerializer.perform_caching && opts[:cache]
+    #     Cacher.call(self, opts, value)
+    #   else
+    #     send_to_serializer(serializer, value)
+    #   end
+    # end
 
 
   end
